Textures added, Door bug fixed, Pause added
Blocks now have texture support, .map changed to allow for this and some temporary textures have been added.  fixed bug when going through doors (caused by using a wrong index that happened to be right most of the time). Changed esc form going to the main menu to pausing the game (intend to add pause menu) and added overlay so buttons can be drawn over the game in the background (might change). Start level can now be set in settings.